\name{as_factor_column}
\alias{as_factor_column}
\title{Convert a column to factor}
\usage{
as_factor_column(df, col)
}
\arguments{
  \item{df}{A data frame}
  \item{col}{A string with column name to convert}
}
\value{
Returns the same data frame with specified column converted to a factor.
}
\description{
A function to convert a character column in a data frame to factor.
}
